Fixed: Static Scroll bar height.
If you enabled an element for slimScroll then added content to that element scrolling would not work as the scrolling is determined from the height of the scrollbar which was set when initially setting up the pane. This fixes the problem by updating the scrollbar height each time it is shown.