Update jquery.slimscroll.js
Fixed the bug:
When user sets "allowPageScroll" to true, the parent should only scroll when the scroll in on the boundary. But when the scroll touched one of the boundaries, the parent started to scroll even when the scroll in not at boundaries

Solution:
The variable "releaseScroll", one set to true, was never set to false. So I set it to false if the scroll is not at the end of the page.